+++ lib/graphql_rails/types/input_object_type.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+require 'graphql_rails/types/argument_type'
+
+module GraphqlRails
+  module Types
+    # Base graphql type class for all GraphqlRails models
+    class InputObjectType < GraphQL::Schema::InputObject
+      argument_class(GraphqlRails::Types::ArgumentType)
+
+      def self.inspect
+        "#{GraphQL::Schema::InputObject}(#{graphql_name})"
+      end
+    end
+  end
+end